multy player test

+++ include/gameLayer/threadStuff.h
@@ -13,7 +13,7 @@ struct Task
 
 	glm::ivec3 pos = {};
 	int type = 0;
-	int blockType = 0;
+	uint16_t blockType = 0;
 
 };
